# contract/openapi.yml
openapi: 3.0.0
info:
  title: IT Corpo API
  version: 2.0.0
  description: >
    API for managing IT corporation resources.

    For full documentation, please visit [the readme](https://github.com/ducin-public/itcorpo-api).
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /offices/amenities:
    get:
      summary: Retrieve list of possible office amenities
      description: Returns an array of office amenity objects that can be assigned to offices
      operationId: getOfficeAmenities
      tags:
        - Offices
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfficeAmenity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /offices/amenities/count:
    get:
      summary: Get total number of office amenities
      operationId: getOfficeAmenitiesCount
      tags:
        - Offices
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 15
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /offices:
    get:
      summary: List all offices
      operationId: getOffices
      tags:
        - Offices
        - Search
      parameters:
        - $ref: '#/components/parameters/OfficePhrase'
        - $ref: '#/components/parameters/OfficeCountries'
        - $ref: '#/components/parameters/OfficeAmenities'
        - $ref: '#/components/parameters/OfficeAmenitiesFiltering'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Office'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    post:
      summary: Create a new office
      operationId: createOffice
      tags:
        - Offices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeInput'
      responses:
        '201':
          description: Office created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '400':
          description: >
            Invalid office input request body
            @see {@link OfficeInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Office with this code already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid office configuration (e.g. invalid country code, unknown amenity)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /offices/count:
    get:
      summary: Get total number of offices
      operationId: getOfficesCount
      tags:
        - Offices
        - Search
      parameters:
        - $ref: '#/components/parameters/OfficePhrase'
        - $ref: '#/components/parameters/OfficeCountries'
        - $ref: '#/components/parameters/OfficeAmenities'
        - $ref: '#/components/parameters/OfficeAmenitiesFiltering'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 25
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /offices/{officeCode}:
    get:
      summary: Get office by office code
      operationId: getOfficeByCode
      tags:
        - Offices
      parameters:
        - name: officeCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '404':
          description: Office not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    
    put:
      summary: Update office
      operationId: updateOffice
      tags:
        - Offices
      parameters:
        - name: officeCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeInput'
      responses:
        '200':
          description: Office updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '400':
          description: >
            Invalid office input request body
            @see {@link OfficeInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Office not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Office code already taken by another office
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid office configuration (e.g. invalid country code, unknown amenity)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      summary: Delete office
      operationId: deleteOffice
      tags:
        - Offices
      parameters:
        - name: officeCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Office deleted successfully
        '404':
          description: Office not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete office that has assigned employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  responses:
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: The service is temporarily unavailable. Maybe there is maintenance?
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Money:
      type: number
      description: Monetary value in EUR
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: An application-level error code (**not** HTTP status code)
        message:
          type: string
          description: Text description of the error that has occurred
      required:
        - message
...
    OfficeAmenity:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          description: Unique code identifier of the amenity
        name:
          type: string
          description: Display name of the amenity
      example:
        code: "parking"
        name: "PARKING"
    Office:
      type: object
      required:
        - code
        - country
        - city
        - address
        - capacity
        - estate
        - amenities
        - monthlyRental
      properties:
        code:
          type: string
        country:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
          minimum: 1
        monthlyRental:
          $ref: '#/components/schemas/Money'
        estate:
          type: object
          required:
            - owner
            - phone
            - account
          properties:
            owner:
              type: string
            phone:
              type: string
            account:
              type: string
        amenities:
          type: array
          items:
            type: string
        imgURL:
          type: string
      example:
        country: "Netherlands"
        city: "Amsterdam" 
        address: "Damrak 81"
        capacity: 150
        monthlyRental: 10000
        estate:
          owner: "B2C Estates and Sons"
          phone: "(7364) 079343"
          account: "NL86 AMUJ 9303 4156 60"
        amenities:
          - code: "OUTDOOR_SEATING"
            name: "Outdoor seating"
          - code: "OPEN_SPACE" 
            name: "Open space"
        imgURL: "amsterdam-6Uf6-XCKJ-qTKq-ISt2-B3SE.jpg"
    OfficeInput:
      type: object
      required:
        - code
        - country
        - city
        - address
        - capacity
        - estate
        - amenities
        - monthlyRental
      properties:
        code:
          type: string
        country:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
          minimum: 1
        monthlyRental:
          $ref: '#/components/schemas/Money'
        estate:
          type: object
          required:
            - owner
            - phone
            - account
          properties:
            owner:
              type: string
            phone:
              type: string
            account:
              type: string
        amenities:
          type: array
          items:
            $ref: '#/components/schemas/OfficeAmenity'
        imgURL:
          type: string
...
  parameters:
...
    OfficePhrase:
      name: phrase
      in: query
      required: false
      description: Full text search across office fields
      schema:
        type: string
        description: Full text search across country, city, address and estate owner fields
        example: "Amsterdam"
    OfficeCountries:
      name: countries
      in: query
      required: false
      description: Filter offices by country codes
      schema:
        type: string
        description: Comma-separated list of country codes to filter by
        example: "PL,US"
    OfficeAmenities:
      name: amenities
      in: query
      required: false
      description: Filter offices by amenities
      schema:
        type: string
        description: Comma-separated list of amenity codes to filter by
        example: "FREE_PARKING,SHOWER"
    OfficeAmenitiesFiltering:
      name: amenitiesFiltering
      in: query
      required: false
      description: How to match office amenities ('ANY' or 'ALL')
      schema:
        type: string
        enum: ["ANY", "ALL"]
        default: "ANY"
        description: If more than one amenity is passed, return either offices with any of the amenities (`ANY`) or with all of them (`ALL`)
        example: "ALL"

    # Project parameters
...